address comments
Rework on response serialization logic, and allow customizable response.
