+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/data/HttpResponseMakerSpec.groovy
@@ -15,8 +15,10 @@ import com.yahoo.bard.webservice.data.metric.MetricColumn
 import com.yahoo.bard.webservice.data.metric.mappers.ResultSetMapper
 import com.yahoo.bard.webservice.druid.model.query.DruidAggregationQuery
 import com.yahoo.bard.webservice.web.DataApiRequest
+import com.yahoo.bard.webservice.web.JsonResponseWriter
 import com.yahoo.bard.webservice.web.PreResponse
 import com.yahoo.bard.webservice.web.ResponseFormatType
+import com.yahoo.bard.webservice.web.ResponseWriter
 import com.yahoo.bard.webservice.web.responseprocessors.ResponseContext
 import com.yahoo.bard.webservice.web.responseprocessors.ResultSetResponseProcessor
 
@@ -45,6 +47,7 @@ class HttpResponseMakerSpec extends Specification {
     ResultSetResponseProcessor resultSetResponseProcessor
     Subject<PreResponse, PreResponse> responseEmitter
     HttpResponseMaker httpResponseMaker
+    ResponseWriter responseWriter
 
     def setup() {
         apiRequest = Mock(DataApiRequest)
@@ -54,8 +57,14 @@ class HttpResponseMakerSpec extends Specification {
         paramMap = Mock(MultivaluedMap)
         DimensionDictionary dimensionDictionary = Mock(DimensionDictionary)
         Dimension dim1 = Mock(Dimension)
+        responseWriter = new JsonResponseWriter(MAPPERS)
 
-        httpResponseChannel = new HttpResponseChannel(Mock(AsyncResponse), apiRequest, new HttpResponseMaker(MAPPERS, dimensionDictionary))
+        httpResponseChannel = new HttpResponseChannel(
+                Mock(AsyncResponse),
+                apiRequest,
+                new HttpResponseMaker(MAPPERS, dimensionDictionary, responseWriter)
+        )
+
         responseEmitter = PublishSubject.create()
         responseEmitter.subscribe(httpResponseChannel)
 
@@ -82,7 +91,7 @@ class HttpResponseMakerSpec extends Specification {
         resultSet = Mock(ResultSet)
         resultSet.getSchema() >> schema
 
-        httpResponseMaker = new HttpResponseMaker(MAPPERS, dimensionDictionary)
+        httpResponseMaker = new HttpResponseMaker(MAPPERS, dimensionDictionary, responseWriter)
         resultSetResponseProcessor = new ResultSetResponseProcessor(
                 apiRequest,
                 responseEmitter,
